requirements.txt
bitsandbytes==0.43.0
clip==0.2.0
cog==0.9.4
deepspeed==0.14.0
editdistance==0.8.1
fastapi==0.110.0
fire==0.6.0
flash_attn==2.5.6
gradio==4.21.0
huggingface_hub==0.21.4
numpy==1.26.4
openai==0.27.10
packaging==24.0
pandas==2.2.1
peft==0.9.0
Pillow==10.2.0
protobuf==4.25.3
pycocoevalcap==1.2
ratelimit==2.2.1
ray==2.9.3
Requests==2.31.0
sglang==0.1.13
shortuuid==1.0.13
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
uvicorn==0.28.0
wget==3.2
xformers==0.0.24
